+++ src/directives/formly-field.test.js
@@ -1,15 +1,18 @@
 const sinon = require('sinon');
+const apiCheck = require('api-check');
 module.exports = ngModule => {
   describe('formly-field', function() {
-    var $compile;
+    let $compile, scope, formlyConfig;
+    let template = '<formly-form form="theForm" model="model" fields="fields"></formly-form>';
     beforeEach(window.module(ngModule.name));
-    beforeEach(inject(function(_$compile_) {
+    beforeEach(inject((_$compile_, $rootScope, _formlyConfig_) => {
       $compile = _$compile_;
+      scope = $rootScope.$new();
+      formlyConfig = _formlyConfig_;
     }));
 
     describe('with template wrapper', function() {
-      var scope, template;
-      beforeEach(inject(function(formlyConfig, $rootScope) {
+      beforeEach(() => {
         formlyConfig.setWrapper([
           {
             types: 'text',
@@ -35,10 +38,8 @@ module.exports = ngModule => {
         formlyConfig.setType({
           name: 'text', template: `<input name="{{id}}" ng-model="model[options.key]" />`
         });
-        scope = $rootScope.$new();
         scope.model = {};
-        template = '<formly-form form="theForm" model="model" fields="fields"></formly-form>';
-      }));
+      });
 
       it('should take the entire wrapper, not just the contents of the wrapper', function() {
         scope.fields = [
@@ -50,8 +51,7 @@ module.exports = ngModule => {
             }
           }
         ];
-        var el = $compile(template)(scope);
-        scope.$digest();
+        const el = compileAndDigest();
         expect(el[0].querySelector('.my-template-wrapper')).to.exist;
       });
 
@@ -66,8 +66,7 @@ module.exports = ngModule => {
             }
           }
         ];
-        var el = $compile(template)(scope);
-        scope.$digest();
+        const el = compileAndDigest();
         var outerEl = el[0].querySelector('.my-other-template-wrapper');
         expect(outerEl).to.exist;
         expect(outerEl.querySelector('.my-template-wrapper')).to.exist;
@@ -84,8 +83,7 @@ module.exports = ngModule => {
             }
           }
         ];
-        var el = $compile(template)(scope);
-        scope.$digest();
+        const el = compileAndDigest();
         expect(el[0].querySelector('.my-template-wrapper')).to.not.exist;
       });
 
@@ -93,19 +91,15 @@ module.exports = ngModule => {
 
 
     describe('api check', () => {
-      let scope, $compile;
-      let template = '<formly-form form="theForm" model="model" fields="fields"></formly-form>';
       let validateOptions;
-      beforeEach(inject((formlyConfig, $rootScope, _$compile_)  => {
-        $compile = _$compile_;
+      beforeEach(() => {
         validateOptions = sinon.spy();
         formlyConfig.setType({
           name: 'text', template: `<input name="{{id}}" ng-model="model[options.key]" />`,
           validateOptions
         });
-        scope = $rootScope.$new();
         scope.model = {};
-      }));
+      });
 
       it('should throw an error when a field has extra properties', () => {
         scope.fields = [
@@ -114,26 +108,21 @@ module.exports = ngModule => {
             extraProp: 'whatever'
           }
         ];
-        $compile(template)(scope);
 
-        expect(() => scope.$digest()).to.throw(/extra.*properties.*extraProp/);
+
+        expect(() => compileAndDigest()).to.throw(/extra.*properties.*extraProp/);
       });
 
       it(`should invoke the validateOptions property of the type`, () => {
         const field = {type: 'text'};
         scope.fields = [field];
-        $compile(template)(scope);
-        scope.$digest();
+        compileAndDigest();
         expect(validateOptions).to.have.been.calledWith(field);
       });
     });
 
     describe('default type options', () => {
-      var scope, $compile;
-      var template = '<formly-form form="theForm" model="model" fields="fields"></formly-form>';
-      beforeEach(inject((_$compile_, formlyConfig, $rootScope) => {
-        $compile = _$compile_;
-        scope = $rootScope.$new();
+      beforeEach(() => {
         scope.model = {};
         formlyConfig.setType({
           name: 'ipAddress', template: '<input name="{{id}}" ng-model="model[options.key]" />',
@@ -188,13 +177,12 @@ module.exports = ngModule => {
             }
           }
         });
-      }));
+      });
 
       it('should default to the ipAddress type options', () => {
         var field = {type: 'ipAddress'};
         scope.fields = [field];
-        $compile(template)(scope);
-        scope.$digest();
+        compileAndDigest();
         expect(field.data.usingDefaultOptions).to.be.true;
         expect(field.validators.ipAddress).to.be.a('function');
       });
@@ -204,8 +192,7 @@ module.exports = ngModule => {
           type: 'text', optionsTypes: ['phone', 'required'], templateOptions: {myChange: 'model.otherThing = true'}
         };
         scope.fields = [field];
-        const el = $compile(template)(scope);
-        scope.$digest();
+        const el = compileAndDigest();
         const input = el.find('input');
         expect(field.data.hasPropertiesFromTextType).to.be.true;
         expect(input.attr('ng-pattern')).to.equal('/overwriting stuff is fun for tests/');
@@ -220,10 +207,9 @@ module.exports = ngModule => {
 
       function testTemplateManipulators(isPre) {
         describe(isPre ? 'preWrapper' : 'postWrapper', () => {
-          var manipulators, scope;
-          var formTemplate = '<formly-form form="theForm" model="model" fields="fields"></formly-form>';
+          var manipulators;
           var textTemplate = '<input class="text-template" name="{{id}}" ng-model="model[options.key]">';
-          beforeEach(inject((formlyConfig, $rootScope) => {
+          beforeEach(() => {
             manipulators = formlyConfig.templateManipulators[isPre ? 'preWrapper' : 'postWrapper'];
             formlyConfig.setWrapper([
               {
@@ -234,12 +220,11 @@ module.exports = ngModule => {
             formlyConfig.setType({
               name: 'text', template: textTemplate
             });
-            scope = $rootScope.$new();
             scope.model = {};
             scope.fields = [
               {type: 'text'}
             ];
-          }));
+          });
 
           var when = isPre ? 'before' : 'after';
 
@@ -274,7 +259,7 @@ module.exports = ngModule => {
               expect(templateToManipulate).to.contain('manipulated');
               return angular.element(templateToManipulate).addClass('manipulated-twice');
             });
-            var el = $compile(angular.element(formTemplate))(scope);
+            var el = $compile(angular.element(template))(scope);
             scope.$digest();
             expect(el[0].querySelector('.manipulated')).to.exist;
             expect(el[0].querySelector('.manipulated-twice')).to.exist;
@@ -288,11 +273,8 @@ module.exports = ngModule => {
     });
 
     describe('type controllers and link functions', () => {
-      var scope, $compile, controllerFn, linkFn;
-      var template = '<formly-form form="theForm" model="model" fields="fields"></formly-form>';
-      beforeEach(inject((formlyConfig, $rootScope, _$compile_) => {
-        $compile = _$compile_;
-
+      let controllerFn, linkFn;
+      beforeEach(() => {
         controllerFn = function($scope) {
           $scope.setInTypeController = true;
         };
@@ -309,15 +291,13 @@ module.exports = ngModule => {
           controller: ['$scope', controllerFn],
           link: linkFn
         });
-        scope = $rootScope.$new();
         scope.model = {};
-      }));
+      });
       it('should run the controller function of a type', () => {
         scope.fields = [
           {type: 'text'}
         ];
-        var el = $compile(template)(scope);
-        scope.$digest();
+        const el = compileAndDigest();
         var fieldEl = angular.element(el[0].querySelector('.formly-field'));
         expect(fieldEl.isolateScope().setInTypeController).to.be.true;
       });
@@ -326,8 +306,7 @@ module.exports = ngModule => {
         scope.fields = [
           {type: 'text'}
         ];
-        var el = $compile(template)(scope);
-        scope.$digest();
+        const el = compileAndDigest();
         var fieldEl = angular.element(el[0].querySelector('.formly-field'));
         var fieldScope = fieldEl.isolateScope();
         expect(fieldScope.setInTypeLink).to.be.true;
@@ -339,8 +318,7 @@ module.exports = ngModule => {
           {template: 'sweet mercy', controller: ['$scope', controllerFn]}
         ];
 
-        var el = $compile(template)(scope);
-        scope.$digest();
+        const el = compileAndDigest();
         var fieldEl = angular.element(el[0].querySelector('.formly-field'));
         expect(fieldEl.isolateScope().setInTypeController).to.be.true;
       });
@@ -350,8 +328,7 @@ module.exports = ngModule => {
         scope.fields = [
           {template: 'sweet mercy', link: linkFn}
         ];
-        var el = $compile(template)(scope);
-        scope.$digest();
+        const el = compileAndDigest();
         var fieldEl = angular.element(el[0].querySelector('.formly-field'));
         var fieldScope = fieldEl.isolateScope();
         expect(fieldScope.setInTypeLink).to.be.true;
@@ -359,5 +336,117 @@ module.exports = ngModule => {
       });
     });
 
+    describe(`type apiCheck`, () => {
+      let type = 'input';
+      beforeEach(() => {
+        formlyConfig.setType({
+          name: type,
+          template: '<label>{{to.label}}</label><input class="{{to.className}}" ng-model="model[options.key]" />',
+          apiCheck: apiCheck.shape({
+            templateOptions: apiCheck.shape({
+              label: apiCheck.string,
+              className: apiCheck.string
+            })
+          }),
+          apiCheckInstance: apiCheck({
+            output: {prefix: 'custom-api-check'}
+          })
+        });
+        scope.model = {};
+      });
+
+      it(`should not warn if everything's fine`, () => {
+        scope.fields = [
+          {type, templateOptions: {label: 'string', className: 'string'}}
+        ];
+        shouldNotWarn(compileAndDigest);
+      });
+
+      it(`should warn if everything's not fine`, () => {
+        scope.fields = [
+          {type, templateOptions: {label: 'string'}}
+        ];
+        shouldWarn(/custom-api-check.*?formly-field.*?className/, compileAndDigest);
+      });
+
+      it(`should throw if the apiCheckFunction is set to "throw" and everything's not fine`, () => {
+        formlyConfig.getType(type).apiCheckFunction = 'throw';
+        scope.fields = [
+          {type, templateOptions: {label: 'string'}}
+        ];
+        expect(compileAndDigest).to.throw(/custom-api-check.*?formly-field.*?className/);
+      });
+    });
+
+    describe(`wrapper apiCheck`, () => {
+      const name = 'input';
+      const template = '<input ng-model="model[options.key]" />';
+      const wrapper = name;
+      beforeEach(() => {
+        formlyConfig.setWrapper({
+          name,
+          template:
+            '<div class="to.className"><label>{{to.label}}</label><formly-transclude></formly-transclude></div>',
+          apiCheck: apiCheck.shape({
+            templateOptions: apiCheck.shape({
+              label: apiCheck.string,
+              className: apiCheck.string
+            })
+          }),
+          apiCheckInstance: apiCheck({
+            output: {prefix: 'custom-api-check'}
+          })
+        });
+        scope.model = {};
+      });
+
+      it(`should not warn if everything's fine`, () => {
+        scope.fields = [
+          {template, wrapper, templateOptions: {label: 'string', className: 'string'}}
+        ];
+        shouldNotWarn(compileAndDigest);
+      });
+
+      it(`should warn if everything's not fine`, () => {
+        scope.fields = [
+          {template, wrapper, templateOptions: {label: 'string'}}
+        ];
+        shouldWarn(/custom-api-check.*?formly-field.*?className/, compileAndDigest);
+      });
+
+      it(`should throw if the apiCheckFunction is set to "throw" and everything's not fine`, () => {
+        formlyConfig.getWrapper(name).apiCheckFunction = 'throw';
+        scope.fields = [
+          {template, wrapper, templateOptions: {label: 'string'}}
+        ];
+        expect(compileAndDigest).to.throw(/custom-api-check.*?formly-field.*?className/);
+      });
+    });
+
+    function compileAndDigest() {
+      const el = $compile(template)(scope);
+      scope.$digest();
+      return el;
+    }
+
+    function shouldWarn(match, test) {
+      var originalWarn = console.warn;
+      var calledArgs;
+      console.warn = function() {
+        calledArgs = arguments;
+      };
+      test();
+      expect(calledArgs[0]).to.match(match);
+      console.warn = originalWarn;
+    }
+
+    function shouldNotWarn(test) {
+      var originalWarn = console.warn;
+      var callCount = 0;
+      console.warn = () => callCount++;
+      test();
+      expect(callCount).to.equal(0);
+      console.warn = originalWarn;
+    }
   });
 };