+++ src/index.common.js
@@ -1,6 +1,13 @@
-var ngModuleName = 'formly';
-var angular = require('./angular-fix');
-var ngModule = angular.module(ngModuleName, []);
+const apiCheck = require('api-check');
+if (!apiCheck) {
+  throw new Error(
+    'angular-formly requires the library apiCheck.js! Please include it! ' +
+      require('./other/docsBaseUrl') + 'apicheckjs-dependency-required'
+  );
+}
+const ngModuleName = 'formly';
+const angular = require('./angular-fix');
+const ngModule = angular.module(ngModuleName, []);
 
 require('./providers')(ngModule);
 require('./services')(ngModule);