+++ src/other/docsBaseUrl.js
@@ -1,0 +1,1 @@
+export default `https://github.com/formly-js/angular-formly/blob/${VERSION}/other/ERRORS_AND_WARNINGS.md#`;